Ejercicio 2 Resuelto y funcional. Se agrega Clase util al ejercicio 02
